requirements.txt
# Adding peep so we can track it on requires.io. For packages
# installation we use peep.py shipped in ./bin
# sha256: 4bmcawK_tCp_yJB624zuXkZLRPiAM1kSArLlYFTjZQU
peep==2.4

# sha256: c1CauJkYk_5byimUihn9wtv5xN1KDpiqnFgDvveKsiY
# sha256: zpqSSG7aW_PZ5Ah_gCSSsVtxsGuJTWgM-Wb1wnlD5TI
# sha256: m7S7kSEJqteivOjiKGp9mgnrzKrv5Br4VYZrZgcVkGA
peepin==0.11
# sha256: agPOL-r90ZOguoom29l3PnV9Ll1eeTOmLqwSmBO9OBo
# sha256: BmutQstMZpROfvz3ME09F_ew6yIuU5WM3YZkINLotBI
Django==1.8

# sha256: 8uJz7TSsu1YJYtXPEpF5NtjfAil98JvTCJuFRtRYQTg
# sha256: ygF2j97N4TQwHzFwdDIm9g7f9cOTXxJDc3jr2RFQY1M
dj-database-url==0.3.0
# sha256: wAr-yzAqmaT4PeybBVxNHMGWkm1iyNsBXWhDLfgRjKg
psycopg2==2.6

# sha256: ufv3R_qOcR0zDy9Da5suz17tnrZ6Y3yoGrzRHCq-3sg
python-decouple==2.3

# sha256: LiSsXQBNtXFJdqBKwOgMbfbkfpjDVMssDYL4h51Pj9s
Jinja2==2.7.3

# sha256: pOwa_1m5WhS0XrLiN2GgF56YMZ2lp-t2tW6ozce4ccM
MarkupSafe==0.23

# sha256: pRec3pItK04EXuWxHocyPud9NjrkApT8glLyXWoOrwY
# sha256: i8g1CCiCrZoBLNeQxGABHk2WvzUS2YoE09q75FOToIk
gunicorn==19.3.0

# sha256: Pk2AGZlglZuCiVGqJMv6o2zr7RSdzXKMEehVeYsV-HA
# sha256: 2slBnbPs4nu1PHQzJD_CLtQs9o3p1sQSk5Dh2a7-YxA
whitenoise==1.0.6

# sha256: 0yV2uq13_dnmhZvpfCtqPBem9IQ2Q4nbcx8Gl7ehlmE
django-csp==2.0.3

# sha256: -KgTrxWsXmi1LfwqCqpBscm1z6jhX6nPdUg390r3TXQ
# sha256: COjgErsGI3rqV4HBr97vAFkNTOvbDAmEOfjy6zBGMxs
# sha256: L-dKvRttGPk8EhVYyUw0EziB9z8ewAWAm0qdt1alqR8
# sha256: XnfmgpOHn92ae5_zegzJEjKUsCDLoYmowbaSLNn3A7U
# sha256: NYsFoMJgXBXPgzfRcBbrla_qrQfJAN-cLN524hlKklg
# sha256: N0EG4BzNrKT3--7HFkwBbZJoTwWqOP6x1QfdL7eJ7x8
# sha256: QJE8a_66Hmsa71Nm22p0w2JT5RplYA6Ldwrte_e50j8
# sha256: ZUnxmBG_0UdReuaLnSXYmgIH3TCMejNL1BlRTjaHXCs
# sha256: _mnJFYY_4PJMANYaS_pevQ-gjYWds6HPyz-yDxpn3UU
# sha256: PhW0FsmiA5waUSCLLNO7T_15bNGeYBsdJlevy3fD3JA
# sha256: c9rw60StkprVaXKtMJOMsjG3JbJ9_xBbcWjntEkeBFI
# sha256: k_gRIthpg8uwVs0Syn3jnXYLbz4wYOQ-_1xeIDRRb6I
pytz==2015.2

# sha256: J7kqegDxYuebWJb-nliXGbmF6c_fRFgmzBiFYVnObB4
newrelic==2.40.0.34

# sha256: wEJTewVQLw7kQNlQ1fONWjwROwPMpHPDnqkAAb2ayFc
# sha256: 0mqdAlLan4Mbgne4JJYbsQsfnXn0dS0y5IFLe80eCJs
flake8==2.4.0

# sha256: SqEp342QB7GSv4IBP0FVM5lGUtfKqTDQAmh-tCpsKkE
# sha256: uLfjVjC1U54moZffxgBb6eHpoTVJazd3I6jrwBuby_8
pep8==1.6.2

# sha256: rEVxaVwQzhU2vNuhopS58tPmzJ0OoXG2fVCghkzj4EI
# sha256: P6gKELNtUWhr93RPXcmWIs1cmM6O1kAi5imGiq_Bd2k
pyflakes==0.8.1

# sha256: ic9r9P9kT_i5EuL6fiq7aNV6jn6Rjjk8coQNMNLrzws
# sha256: PYypv2XFAU9GkYBUTR3Vu1ud9wmq1jBPnC5DcK4Ke3w
mccabe==0.3

# sha256: 9WaHQjOU18jXr8lhEuT31usl1zancza3sYqwnpscDNI
# sha256: 4J4tQ_7ka9T7TiHCP-clhdw8NDcHMCf_iXYJ3Vt1dLw
# sha256: hgcBQREBAba1Xm3DGkBGHEN9ORpTN6bl40k1fcGL-7Q
# sha256: OpaC7xSjGTR-0VOuFxRlvHb2KOW1lllrQM8KKeN79wc
ipython==3.0.0

# sha256: DUoz845megyA7BpaDkaP5Y5c3GHLj0B3sQJLhf57cRc
ipdb==0.8

# sha256: Ld8Y2jsGIfpD_uS3KQ2grnibRvuJkniorMzaGVxJeac
# sha256: Gm5RMMK0LS3jAWk8KZ94zEvTUB54thDAjkXvxw4rURQ
Sphinx==1.3.1

# sha256: qKLiOalzG2hcfepiSF3sgc7zYB6-UKd6lalr3yl6-_c
# sha256: mkkMhh9s-WoAUMKaktXR4B7aAq5vUHYK1clqMnzfFOg
sphinx-rtd-theme==0.1.7

# sha256: unY7cDzPDaegZs8EGxqbokyQOzNsvl7t3m63L18mzJE
# sha256: svoxRSmeK3IsSrCM_Ok5JkOzx3E_oQtpT90mMVnJr6s
sphinx-autobuild==0.5.0

# sha256: x9txeBCraWX2bIzwOYqYydjfmC2jm0zX8WKRHriVlvo
docutils==0.12

# sha256: CjoiZeHvsN76ci0fQpcw3J35da3DxgrH0rQyvfA8BB0
# sha256: Xe0vqQlP19_rPakmNkCf1wKg0H1gYoNQTX7gRAHO5cs
# sha256: cyCRkITm2sj0VAY4pGRHo71zD8oXKvwX0sA-7SLPT1E
Pygments==2.0.2

# sha256: 3w1cLeq8MzdfKIx6tUYuBsbs2CTs-ie_qLzV5Mz7Nyo
django-jinja==1.3.3

# sha256: F2ceHw8e9cEkIpH3EJVrqURr-AT_kH6rH-UzAxuUEbU
django-auth-ldap==1.2.6

# sha256: 6a8WD_NhGTt5vv2eFmNv4cImZybnOMyjIQwoxRmBFRY
django-generic-json-views==0.6.5b

# saha256: Bld7cwKKuPw-lb54UkXDNeDOULezCB-HczyVAxzxn48
# phonenumbers==7.0.4

# sha256: 87I68pDQ-ACUXETgc3vTLz7jj2PnrXB2RQd4_mgqUv4
django-phonenumber-field==0.7.2

# sha256: pLr479ZvCvUyhqconkn0bYDE0HCcMknunIdZdmZkIF8
djangorestframework==3.1.1

# sha256: 7hfQ19wJHmRd1IMCouITAcxo8YhQXCBp2GNflFVBcL8
markdown==2.6.2

# sha256: 5dkCOGALbqnyFYiARBpH5fPwOZzvtT74QEDqxX1yVWM
django-filter==0.9.2

# sha256: ByQht478gXZBLpRL8auUDiO2WLInD0Xw8O-aw61q5gU
django-pipeline==1.5.0

# sha256: MVg9nhJ2eR2EqgFv3IeR_RhSisV5fbrhGLMOMuApDE8
django-jsonview==0.4.3
